Airhockey builds and starts up now, but is not functional.  Audio plays, but nothing renders and you can't interact with the window.  Only seems to work launching from Xcode

+++ Engine/src/Audio/icSoundDevice.cpp
@@ -161,7 +161,7 @@ icSoundI* icSoundDeviceI::GetFreeSound(void)
 **/
 ICRESULT icSoundDeviceI::FreeSound(icSoundI* pSound)
 {
-    size_t i = icMinHeapFind(m_Playing,m_NumPlaying,pSound);
+    int i = icMinHeapFind(m_Playing,m_NumPlaying,pSound);
     if (i >= 0)
     {
         icMinHeapDecreaseKey(m_Playing,i,(icSoundI*)0);
